move internal and fix mv
