+++ aeon/classification/convolution_based/_mr_hydra.py
@@ -4,7 +4,10 @@ from sklearn.preprocessing import StandardScaler
 
 from aeon.classification import BaseClassifier
 from aeon.classification.convolution_based._hydra import _SparseScaler
-from aeon.transformations.collection.convolution_based import MultiRocket
+from aeon.transformations.collection.convolution_based import (
+    MultiRocket,
+    MultiRocketMultivariate,
+)
 from aeon.transformations.collection.convolution_based._hydra import HydraTransformer
 
 
@@ -14,7 +17,7 @@ class MultiRocketHydraClassifier(BaseClassifier):
     _tags = {
         "capability:multivariate": True,
         "capability:multithreading": True,
-        "classifier_type": "convolution",
+        "algorithm_type": "convolution",
         "python_dependencies": "torch",
     }
 
@@ -38,9 +41,16 @@ class MultiRocketHydraClassifier(BaseClassifier):
         self._scale_hydra = _SparseScaler()
         Xt_hydra = self._scale_hydra.fit_transform(Xt_hydra)
 
-        self._transform_multirocket = MultiRocket(
-            n_jobs=self.n_jobs,
-            random_state=self.random_state,
+        self._transform_multirocket = (
+            MultiRocket(
+                n_jobs=self.n_jobs,
+                random_state=self.random_state,
+            )
+            if X.shape[1] == 1
+            else MultiRocketMultivariate(
+                n_jobs=self.n_jobs,
+                random_state=self.random_state,
+            )
         )
         Xt_multirocket = self._transform_multirocket.fit_transform(X)
